requirements.txt
torch==1.7.1
tqdm
json_tricks
opencv_python
prettytable
scipy
torchvision
numpy
matplotlib
easydict
PyYAML
tensorboardX
pycocotools
